requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
flask==2.3.3
flask-socketio==5.3.6
gunicorn==21.2.0
numpy==1.24.3
pandas==2.0.3
